// server/src/types/playerStats.ts
export interface SeasonStatsEntry {
  season: number;
  games: number;
  completions: number;
  attempts: number;
  passingYards: number;
  passingTds: number;
  interceptions: number;
  carries: number;
  rushingYards: number;
  rushingTds: number;
  receptions: number;
  targets: number;
  receivingYards: number;
  receivingTds: number;
  fantasyPoints: number;
  fantasyPointsPpr: number;
}